Squash cast and unused var warnings

+++ src/drivers/cs_Storage.cpp
@@ -452,7 +452,7 @@ void Storage::getInt8(int32_t value, int8_t* target, int8_t default_value) {
 
 	// check if last byte is FF which means that memory is unassigned
 	// and value has to be ignored
-	if (value == UINT32_MAX) {
+	if ((uint32_t)value == UINT32_MAX) {
 #ifdef PRINT_ITEMS
 		LOGd(FMT_USE_DEFAULT_VALUE);
 #endif
